Add generateEmptyTx config param
if generateEmptyTx is true when user want to attach stage
we generate an empty light transaction and attach stage
if not we return error code
