+++ env.js.example
@@ -14,7 +14,8 @@ let env = {
     host: '<DATABASE_HOST>',
     dialect: 'postgres',
     logging: false
-  }
+  },
+  generateEmptyTx: true
 };
 
 if (!argv.hasOwnProperty('migrations-path')) {
